fourth commit

+++ consultant_app/views.py
@@ -17,6 +17,9 @@ from django.http import Http404
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from rest_framework.pagination import PageNumberPagination
 from rest_framework import permissions
+from rest_framework.parsers import MultiPartParser, FormParser
+import os
+# from os.path import abspath, dirname
 
 class UserPermissionsObj(permissions.BasePermission):
 
@@ -76,7 +79,7 @@ class LargeResultsSetPagination(PageNumberPagination):
 class SupporterPanelViewSet(viewsets.ReadOnlyModelViewSet):
     # authentication_classes = (TokenAuthentication,)
     # permission_classes = (IsAuthenticated,)
-    queryset = User.objects.filter(role='supporter')
+    queryset = User.objects.filter(role='supporter').filter(is_active=True)
     serializer_class = SupporterPanelSerializer
 
     @detail_route(methods=["GET", "POST"])
@@ -87,10 +90,17 @@ class SupporterPanelViewSet(viewsets.ReadOnlyModelViewSet):
             if user.is_superuser == True:
                 return Response("User Is Admin So Cannot Be Deleted", status=status.HTTP_406_NOT_ACCEPTABLE)
             else:
-                user.delete()
-                data=User.objects.all()
-                serializers=SupporterDetailSerializer(data,many=True)
-                return Response(serializers.data, status=status.HTTP_200_OK)
+                if user.role == 'consultant':
+                    user.delete()
+                    data=User.objects.filter(role='consultant')
+                    serializers=CreateConsultantSerializer(data,many=True)
+                    return Response(serializers.data, status=status.HTTP_200_OK)
+                else:
+                    if user.role== 'supporter':
+                        user.delete()
+                        data = User.objects.filter(role='supporter')
+                        serializers = SupporterDetailSerializer(data, many=True)
+                        return Response(serializers.data, status=status.HTTP_200_OK)
         except User.DoesNotExist:
             raise Http404("No User matches the given query.")
 
@@ -118,9 +128,35 @@ class SupporterPanelViewSet(viewsets.ReadOnlyModelViewSet):
         except To_do_list.DoesNotExist:
             raise Http404("No to_do_list matches the given query.")
 
+    @detail_route(methods=["GET", "POST"])
+    def delete_notification(self, request, pk=None):
+        try:
+            notification = Notification.objects.filter(recipient_id=request.user.id).get(id=pk)
+            notification.delete()
+            data = Notification.objects.filter(recipient_id=request.user.id)
+            serializers = NotificationSerializer(data, many=True)
+            return Response(serializers.data, status=status.HTTP_200_OK)
+
+        except Notification.DoesNotExist:
+            raise Http404("No notification matches the given query.")
+
+    @detail_route(methods=["GET", "POST"])
+    def active_user(self, request, pk=None):
+        try:
+            user = User.objects.get(id=pk)
+            user.is_active=True
+            user.save()
+            data = User.objects.filter(id=pk)
+            serializers = SupporterDetailSerializer(data, many=True)
+            return Response(serializers.data,status=status.HTTP_200_OK)
+        except User.DoesNotExist:
+            raise Http404("No user matches the given query.")
+
+
 class AddConsultantViewSet(viewsets.ModelViewSet):
     # authentication_classes = (TokenAuthentication,)
     # permission_classes = (IsAdminUser,)
+    # parser_classes = (MultiPartParser, FormParser, )
     queryset = User.objects.filter(role='consultant')
     serializer_class = CreateConsultantSerializer
     filter_backends = (filters.DjangoFilterBackend, filters.SearchFilter)
@@ -131,7 +167,9 @@ class AddConsultantViewSet(viewsets.ModelViewSet):
         serializer = CreateConsultantSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
-            return Response({'status': 'Success'})
+            value=serializer.data['resume']
+            # new_dict.update({'resume_name':os.path.basename(serializer.data['resume'])})
+            return Response({'message':'success'})
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
@@ -139,9 +177,9 @@ class AddConsultantViewSet(viewsets.ModelViewSet):
 class AddSupporterViewSet(viewsets.ModelViewSet):
     # authentication_classes = (TokenAuthentication,)
     # permission_classes = (IsAdminUser,)
+    # parser_classes = (MultiPartParser, FormParser, )
     queryset = User.objects.filter(id=0)
     serializer_class = CreateSupporterSerializer
-
 
     # def create(self, request, *args, **kwargs):
     #     serializer = self.get_serializer(data=request.data)
@@ -181,17 +219,18 @@ class AddProjectViewSet(viewsets.ModelViewSet):
 
     @detail_route(methods=['GET', 'POST'])
     def get_comments(self, request, pk=None):
-        queryset = Project.objects.filter(id=pk)
-        serializers = ProjectInfoSerializer(queryset, many=True)
+        data = Project.objects.get(id=pk)
+        serializers = ProjectInfoSerializer(data)
         data=serializers.data
-        return Response(data.pop(0))
+        return Response(data)
 
 class SupporterDetailViewset(viewsets.ModelViewSet):
     # authentication_classes = (TokenAuthentication,)
     # permission_classes = (IsAuthenticated,)
-    queryset = User.objects.filter(role='supporter')
+    # parser_classes = (MultiPartParser, FormParser, )
+    queryset = User.objects.filter(role='supporter').filter(is_active=True)
     serializer_class = SupporterDetailSerializer
-    permission_classes=(UserPermissionsObj,)
+    # permission_classes=(UserPermissionsObj,)
     filter_backends = (filters.DjangoFilterBackend, filters.SearchFilter)
     search_fields = ('first_name',)
     filter_fields = ('first_name',)
@@ -249,7 +288,7 @@ class UserViewSet(viewsets.ViewSet):
     def list(self, request):
         user= Token.objects.get(key=request.auth)
         queryset= User.objects.get(username=user.user)
-        serializer=UserSerializer(queryset)
+        serializer=UserSerializer(queryset,context={'request': request})
         return Response(serializer.data)
